move 'filter' tests to separate file
